命名错误

+++ build/afterBuild.sh
@@ -26,7 +26,11 @@ GitAddSite1_0 () {
     git rm -rf ./static
     git rm ./index.html
     cp ../src/index.html ./index.html
-    cp ../site/app.js ./static/app.js
+    if [ ! -f "../site/app.js"]; then
+        exit 1
+    else
+        cp ../site/app.js ./static/app.js
+    fi
     cp ../build/empty_circle.yml ./circle.yml
     sed -i '15i\    <script type="text/javascript" src="./static/app.js"></script>' ./index.html
     mkdir static
